# api/openApi/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger API documentation for test task",
        "contact": {
            "name": "Dmytro",
            "email": "dm.kopylets@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/process-payload": {
            "post": {
                "tags": [
                    "Price"
                ],
                "summary": "main action",
                "operationId": "processPayload",
                "requestBody": {
                    "description": "Client side request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "type": "object",
                                        "additionalProperties": {
                                            "type": "object",
                                            "properties": {
                                                "currency": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "number"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    },
                                    "checkoutCurrency": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Price"
        }
    ]
}